Updated 2024-10-08 09:59 PDT
signalviewer: The panning behavior when pressing up or down arrow on the PSD window is changed to moving one channel up or down (and pan if necessary). This way the user can browse the PSD between channels more easily.

convert_timeseries_to_rawdata: The opposite of convert_rawdata_to_timeseries, is added.